+++ SqlTark/Connection/SqlServerConnection.php
@@ -1,0 +1,70 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SqlTark\Connection;
+
+use SqlTark\Connection\AbstractConnection;
+
+class SqlServerConnection extends AbstractConnection
+{
+    /**
+     * {@inheritdoc}
+     */
+    protected string $driver = 'sqlsrv';
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function createDSN(): string
+    {
+        $host = $this->config->getHost();
+        $port = $this->config->getPort();
+        $database = $this->config->getDatabase();
+
+        $dsn = "sqlsrv:Server={$host}";
+        if(!empty($port)) $dsn = "{$dsn},{$port}";
+        $dsn = "{$dsn};Database={$database}";
+        return $dsn;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function transaction(): bool
+    {
+        if($this->transactionCount++ === 0) {
+            return $this->getPDO()->beginTransaction();
+        }
+
+        $this->getPDO()->exec("SAVE TRANSACTION __trx_{$this->transactionCount}__");
+        return $this->transactionCount >= 0;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function commit(): bool
+    {
+        if(--$this->transactionCount === 0) {
+            return $this->getPDO()->commit();
+        }
+
+        return $this->transactionCount >= 0;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function rollback(): bool
+    {
+        if($this->transactionCount > 1) {
+            $this->getPDO()->exec("ROLLBACK TRANSACTION __trx_{$this->transactionCount}__");
+            $this->transactionCount--;
+            return true;
+        }
+
+        $this->transactionCount--;
+        return $this->getPDO()->rollBack();
+    }
+}