+++ CHANGELOG.md
@@ -5,15 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.2] - 2023-03-13
+
+### Added
+- Added `SqlTark\Connection\*` classes
+- Add implementation of `MySqlConnection`
+- Add implementation of `SqlServerConnection`
+- Add implementation of `SqlServerDbLibConnection`
+
+### Changed
+- Change parameter type of `limit` and `offset` from `int` to `scalar`
+
 ## [0.1.1] - 2023-03-12
 
-## Added
+### Added
 - Add support query as parameter for condition `in` clauses
 - Add `compile` method in `SqlTark\Query`
 - Add test cases for `SqlTark\Compiler\MySqlCompiler`
 - Add `flatten` functin in `SqlTark\Utilities\Helper`
 
-## Changed
+### Changed
 - Change implementation of `clearComponents` method in `SqlTark\Query\AbstractQuery`
 
 ### Fixed